Implementando cadastro de moedas
- Agora é possível cadastrar e apagar novas moedas no sistema;
- Algumas alterações nos services de algumas entidades.

+++ numismaster/src/main/java/com/numismaster/service/CoinService.java
@@ -1,7 +1,12 @@
 package com.numismaster.service;
 
+import java.util.List;
+
 import com.numismaster.model.Coin;
 import com.numismaster.repository.CoinRepository;
+
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 
 public class CoinService {
     private CoinRepository coinRepository;
@@ -11,11 +16,45 @@ public class CoinService {
     }
 
     public boolean save(Coin coin){
-        
+        if(findByName(coin.getName()) != null){
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("OPS...");
+            alert.setHeaderText("Moeda duplicada.");
+            alert.setContentText("O nome de moeda informado já está cadastrado no sistema.");
+            alert.showAndWait();
+        } else {
+            if(coin.getId() == 0){
+                if(coinRepository.insert(coin)){
+                    return true;
+                }
+            } else {
+                if(coinRepository.update(coin)){
+                    return true;
+                }
+            }
+        }
         return false;
     }
 
     public boolean delete(Coin coin){
-        return coinRepository.delete(coin.getId());
+        CoinUserService coinUserService = new CoinUserService();
+        if(coinUserService.findUsersByCoin(coin) != null){
+            Alert alert = new Alert(AlertType.WARNING);
+            alert.setTitle("OPS...");
+            alert.setHeaderText("Não é possível excluir a moeda.");
+            alert.setContentText("A moeda está sendo utilizada por algum usuário no sistema.");
+            alert.showAndWait();
+        } else {
+            return coinRepository.delete(coin.getId());
+        }
+        return false;
+    }
+
+    public List<Coin> findAll(){
+        return coinRepository.findAll();
+    }
+
+    public Coin findByName(String name){
+        return coinRepository.findByName(name);
     }
 }
